requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==1.26.4
ultralytics==8.2.28
torch==2.3.0
torchvision==0.18.0